Rename crate to star-constellation
We've decided on the name "Constellation" for this extension of
the STAR threshold aggregtation protocol, so aligning the crate
name with other descriptions.

The name 'constellation' is already used by several other projects
on crates.io, so prepend 'star' to distinguish this implementation
and hopefully avoid conflicts when the code is published there.