ci: switch to dtolnay/rust-toolchain
The `actions-rs` suite of Github Actions scripts is very nice,
but it has been unmaintained for the past couple of years. The
toolchain action is now warning that it's using deprecated
dependencies from github, which may be disabled later this year.

Switch to dtolnay/rust-toolchain to install the rust toolchain,
but stick with actions-rs/clippy and actions-rs/audit for now
since invoking the commands directly doesn't generate the nice
reports.