github-actions: pin third-party action script versions
Use a fixed revision instead of movable tags so we get a
chance to review any changes. Renovate or Dependabot should
prompt to bump these commit ids if there are new releases.

The actions-rs ci suite has been unmaintained for some years,
so those are old, but widely used, revisions.

The rust-toolchain repo defines branches and tags aligned
with each rust release, so one can use `@stable` or `@1.56.1`
to specify action script and toolchain version on one line.
This is incompatible with pinning the action script itself.

The instructions say to use `@master` in combination with
a `with` clause specifying the toolchain version like we do
to allow a matrix expansion. But if I've understood the current
code correctly, that's not necessary because the version is
just passed to `rustup` after some filtering.